refactor: change get current dirname to helpers in main path
