+++ src/main/helpers/get-current-dirname.ts
